move login to ui
